+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/DomainMapping.kt
@@ -163,6 +163,7 @@ internal class DomainMapping(
             name = name ?: "",
             image = image ?: "",
             watcherCount = watcher_count,
+            filterTags = filter_tags.orEmpty(),
             frozen = frozen,
             createdAt = created_at,
             deletedAt = deleted_at,